requirements.txt
django==1.11.1
django-autoslug==1.9.3